requirements.txt
sqlalchemy==2.0.23
alembic==1.13.1
typer==0.9.0
rich==13.7.0
python-frontmatter==1.1.0
sentence-transformers==2.2.2
numpy==1.26.2
anthropic==0.8.1
streamlit==1.28.2
pytest==7.4.3
hypothesis==6.92.1
python-dotenv==1.0.0
pyyaml==6.0.1